refactor: вынес логику работы с файлами в FileManager
- Устранил нарушение принципа единой ответственности (SRP) в классе Manager
- Вынес функционал записи и чтения файлов в отдельный класс FileManager
- Класс Manager теперь занимается только своей основной задачей
